Rename interview command to review
