+++ packages/alfa-cli/src/tsconfig.json
@@ -10,28 +10,28 @@
     "alfa/command/common/answering-oracle.ts",
     "alfa/command/common/question-store.ts",
     "alfa/command/common/recording-oracle.ts",
-    "alfa/command/interview/answer/arguments.ts",
-    "alfa/command/interview/answer/flags.ts",
-    "alfa/command/interview/answer/run.ts",
-    "alfa/command/interview/list/arguments.ts",
-    "alfa/command/interview/list/flags.ts",
-    "alfa/command/interview/list/run.ts",
-    "alfa/command/interview/reset/arguments.ts",
-    "alfa/command/interview/reset/flags.ts",
-    "alfa/command/interview/reset/run.ts",
-    "alfa/command/interview/status/arguments.ts",
-    "alfa/command/interview/status/flags.ts",
-    "alfa/command/interview/status/run.ts",
-    "alfa/command/interview/answer.ts",
-    "alfa/command/interview/arguments.ts",
+    "alfa/command/review/answer/arguments.ts",
+    "alfa/command/review/answer/flags.ts",
+    "alfa/command/review/answer/run.ts",
+    "alfa/command/review/list/arguments.ts",
+    "alfa/command/review/list/flags.ts",
+    "alfa/command/review/list/run.ts",
+    "alfa/command/review/reset/arguments.ts",
+    "alfa/command/review/reset/flags.ts",
+    "alfa/command/review/reset/run.ts",
+    "alfa/command/review/status/arguments.ts",
+    "alfa/command/review/status/flags.ts",
+    "alfa/command/review/status/run.ts",
+    "alfa/command/review/answer.ts",
+    "alfa/command/review/arguments.ts",
     "alfa/command/common/digest.ts",
-    "alfa/command/interview/flags.ts",
-    "alfa/command/interview/list.ts",
-    "alfa/command/interview/question-descriptions.ts",
-    "alfa/command/interview/reset.ts",
-    "alfa/command/interview/run.ts",
-    "alfa/command/interview/status.ts",
-    "alfa/command/interview/utils.ts",
+    "alfa/command/review/flags.ts",
+    "alfa/command/review/list.ts",
+    "alfa/command/review/question-descriptions.ts",
+    "alfa/command/review/reset.ts",
+    "alfa/command/review/run.ts",
+    "alfa/command/review/status.ts",
+    "alfa/command/review/utils.ts",
     "alfa/command/query/arguments.ts",
     "alfa/command/query/flags.ts",
     "alfa/command/query/run.ts",
@@ -39,7 +39,7 @@
     "alfa/command/scrape/flags.ts",
     "alfa/command/scrape/run.ts",
     "alfa/command/audit.ts",
-    "alfa/command/interview.ts",
+    "alfa/command/review.ts",
     "alfa/command/query.ts",
     "alfa/command/scrape.ts",
     "alfa/profiler.ts",
